HELM-373: Refactor dashboard convert, filter panel convert, enable setting dashboard title, remove uid
