+++ src/lib/dashboard-convert/index.ts
@@ -1,0 +1,7 @@
+import { dashboardConvert } from './convert'
+import { getDashboardTitle } from './utils'
+
+export {
+  dashboardConvert,
+  getDashboardTitle
+}
